# main/docs/swagger.yaml
basePath: /
definitions:
...
  dto.LocaleResponse:
    properties:
      en:
        description: 英语
        type: string
      ja:
        description: 日语
        type: string
      ko:
        description: 韩语
        type: string
      my:
        description: 缅甸语
        type: string
      th:
        description: 泰语
        type: string
      tr:
        description: 土耳其语
        type: string
      zh:
        description: 中文
        type: string
      zhtw:
        description: 粤语
        type: string
    type: object
  dto.PageResponse:
    properties:
      page_no:
        description: 当前页码
        type: integer
      page_size:
        description: 每页大小
        type: integer
      total:
        description: 总数
        type: integer
    type: object
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  product_resp.ProductCategory:
    properties:
      children:
        allOf:
        - $ref: '#/definitions/product_resp.ProductCategoryListResp'
        description: 子级类别
      is_special:
        description: 是否特殊类别
        type: boolean
      locale_name:
        allOf:
        - $ref: '#/definitions/dto.LocaleResponse'
        description: 商品类别名称
      parent_uuid:
        description: 父级类别UUID
        type: integer
      uuid:
        description: 商品类别UUID
        type: integer
    type: object
  product_resp.ProductCategoryListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/product_resp.ProductCategory'
        type: array
    type: object
...
  req.CallReq:
    properties:
      call_type:
        description: 呼叫类型1-服务员；2-结账
        enum:
        - 1
        - 2
        type: integer
      desk_uuid:
        description: 桌台ID
        type: integer
    required:
    - call_type
    - desk_uuid
    type: object
...
  req.LoginReq:
    properties:
      brand:
        description: 品牌名称
        type: string
      code:
        description: 验证码
        type: string
      device_id:
        description: 设备ID
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - code
    - device_id
    - password
    - username
    type: object
...
  req.ProcessedCallReq:
    properties:
      uuid:
        type: integer
    required:
    - uuid
    type: object
  req.ProductUuid:
    properties:
      product_uuid:
        description: 送厨商品ID
        type: integer
    type: object
...
  resp.ProductionGroup:
    properties:
      locale_name:
        allOf:
        - $ref: '#/definitions/dto.LocaleResponse'
        description: 序列号
      product_list:
        allOf:
        - $ref: '#/definitions/resp.ProductionList'
        description: 送厨商品列表
    type: object
  resp.ProductionHistory:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.ProductionGroup'
        type: array
    type: object
  resp.ProductionItem:
    properties:
      create_time:
        description: 送厨时间
        type: integer
      locale_name:
        allOf:
        - $ref: '#/definitions/dto.LocaleResponse'
        description: 送厨商品名称
      num:
        description: 送厨商品数量
        type: integer
      product_attribute_names:
        description: 商品属性
        type: string
      serial_no:
        description: 序列号
        type: string
      uuid:
        description: 送厨商品Uuid
        type: integer
    type: object
  resp.ProductionList:
    properties:
      list:
        description: 送厨商品
        items:
          $ref: '#/definitions/resp.ProductionItem'
        type: array
    type: object
  resp.ProductionListWithPagination:
    properties:
      finished_list:
        allOf:
        - $ref: '#/definitions/resp.ProductionList'
        description: 最近三个上菜历史
      list:
        description: 分组列表
        items:
          $ref: '#/definitions/resp.ProductionGroup'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/dto.PageResponse'
        description: 分页信息
    type: object
...
  resp.UnprocessedCallItem:
    properties:
      call_type:
        description: 呼叫类型:呼叫类型(1服务员,2结账)
        type: integer
      desk_no:
        description: 桌台编号
        type: string
      desk_uuid:
        description: 桌台Uuid
        type: integer
      is_send:
        description: 是否已发送：1-是；0-否
        type: integer
      uuid:
        description: 呼叫Uuid
        type: integer
    type: object
  resp.UnprocessedCallList:
    properties:
      list:
        description: 未处理呼叫列表
        items:
          $ref: '#/definitions/resp.UnprocessedCallItem'
        type: array
      meta:
        $ref: '#/definitions/dto.PageResponse'
    type: object
  resp.UnprocessedResp:
    properties:
      abnormal_print_count:
        description: 异常打印数量
        type: integer
      unprocessed_call_count:
        description: 未处理呼叫数量
        type: integer
    type: object
...
host: 127.0.0.1:8081
info:
  contact: {}
  description: 点餐系统服务端接口文档
  title: ttpos-server-go API
  version: "1.0"
paths:
...
  /kitchen/call/processed:
    post:
      consumes:
      - application/json
      description: 处理呼叫
      parameters:
      - description: 处理呼叫参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ProcessedCallReq'
      produces:
      - application/json
      responses: {}
      security:
      - JwtToken: []
      summary: 处理呼叫
      tags:
      - 厨显端.呼叫
  /kitchen/call/unprocessed:
    get:
      consumes:
      - application/json
      description: 获取未处理消息数量
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.UnprocessedResp'
              type: object
      security:
      - JwtToken: []
      summary: 获取未处理消息数量
      tags:
      - 厨显端.呼叫
  /kitchen/call/unprocessed/list:
    get:
      consumes:
      - application/json
      description: 未处理呼叫列表
      parameters:
      - description: 页码
        in: query
        name: page_no
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.UnprocessedCallList'
              type: object
      security:
      - JwtToken: []
      summary: 未处理呼叫列表
      tags:
      - 厨显端.呼叫
...
  /kitchen/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 验证码sign
        in: header
        name: X-SIGN
        required: true
        type: string
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 登录
      tags:
      - 厨显端.认证鉴权
  /kitchen/logout:
    post:
      consumes:
      - application/json
      description: 退出登录
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - JwtToken: []
      summary: 退出登录
      tags:
      - 厨显端.认证鉴权
  /kitchen/product/category/list:
    get:
      consumes:
      - application/json
      description: 获取产品类别列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/product_resp.ProductCategoryListResp'
        "400":
          description: 错误请求
      security:
      - JwtToken: []
      summary: 获取产品类别列表
      tags:
      - 厨显端.产品
  /kitchen/product/finish:
    post:
      consumes:
      - application/json
      description: 制作完成
      parameters:
      - description: 送厨商品Uuid参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ProductUuid'
      produces:
      - application/json
      responses: {}
      security:
      - JwtToken: []
      summary: 制作完成
      tags:
      - 厨显端.产品
  /kitchen/product/history:
    get:
      consumes:
      - application/json
      description: 获取上菜历史
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.ProductionHistory'
              type: object
      security:
      - JwtToken: []
      summary: 获取上菜历史
      tags:
      - 厨显端.产品
  /kitchen/product/list_by_category:
    get:
      consumes:
      - application/json
      description: 根据分类获取送厨商品列表
      parameters:
      - description: 分类Uuid，0-全部
        in: query
        name: category_uuid
        type: integer
      - description: 页码
        in: query
        minimum: 1
        name: page_no
        type: integer
      - description: 每页大小
        in: query
        maximum: 1000
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.ProductionListWithPagination'
              type: object
      security:
      - JwtToken: []
      summary: 根据分类获取送厨商品列表
      tags:
      - 厨显端.产品
  /kitchen/product/list_by_order:
    get:
      consumes:
      - application/json
      description: 根据订单获取送厨商品列表
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: page_no
        type: integer
      - description: 每页大小
        in: query
        maximum: 1000
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.ProductionListWithPagination'
              type: object
      security:
      - JwtToken: []
      summary: 根据订单获取送厨商品列表
      tags:
      - 厨显端.产品
  /kitchen/product/recovery:
    post:
      consumes:
      - application/json
      description: 恢复制作
      parameters:
      - description: 送厨商品Uuid参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ProductUuid'
      produces:
      - application/json
      responses: {}
      security:
      - JwtToken: []
      summary: 恢复制作
      tags:
      - 厨显端.产品
...
  /tablet/call:
    post:
      consumes:
      - application/json
      description: 发起呼叫
      parameters:
      - description: 呼叫请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.CallReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - JwtToken: []
      summary: 发起呼叫
      tags:
      - 平板端.呼叫
...
securityDefinitions:
  JwtToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"